Pin supported clang-format to 10, fix usage
Found clang-format binary was not used in ClangFormat.make which
caused confusion if you had more than one installed and the found
was not the same as your alias.

Version 10 has a slight difference how some formatting is done so
to keep things stable we now use it as our supported one. We also
pick 11 already as the next best thing over 9.