MinGW: Remove static fixed instances from constructors
OptionsClass and RulesClass are used as static globals. Because
static intialization order is undefined the const fixed instances
may not be initialized before these classes.

Uninitialized fixeds evaluate to zero which causes random things
to happen in the engine. RA will be monochrome and some rules have
wrong values.

Alternative fix would be to ensure the pre-defined fixeds are
initialized at compile time as constexpr but it's not a trivial
change.