[RA] Fix potential buffer over- and underflows
Any code processing lists expects the last list item to be EOL but
the list copying function did not ensure the last item is EOL.

Any unit that is not on a mission (yet) has MISSION_NONE set which
is -1 and it is used to directly index the global mission list which
will underflow if it's not checked.

Both of these issues were uncovered by ASAN and kept crashing the
game when we were in an errorneous state caused by incorrect
endianness being used.

One overlapping memcpy was also changed to memmove.