[RA] Fix animation crash on GCC
The game engine depends on fixed heap pool management by returning
nullptr from a fixed heap derived child class when the heap is full.

Watcom and MSVC apparently add nullptr guards when a child class
overloaded new operator fails but GCC does not do this by default
which causes a crash when a base constructor is called with
this pointer being null.

@tomsons26 narrowed down the issue and found the GCC option.

Fixes #54