RawFileClass support for case sensitive filesystems
Always make a defensive copy of the filename and try to resolve it
to an actual file on the disk. Works with directory paths as well.

The implementation is not doing proper bounds checking but I
figure it's better to have something working and clean all string
cat/copy usage later as the games are not very memory safe to
begin with.