[RA] Remove LZO compression support
LZO compression was only used for saves and is only marginally
better for them than LCW. The implementation is not 64-bit safe
and is quite hefty.

Saves are now compressed with LCW which still achieves files that
are almost the same size.

Commented out LCWPipe/LCWStraw next to LZO in saveload.cpp also
indicate the developers were originally planning to use LCW but
for some reason decided to bring in LZO support instead.

We could also make them uncompressed but since we still have a
working compressor that we are required to keep around for other
purposes it fits the bill nicely.