[RA] Refactors Force_CD_Available to add better no-cd support.
Game will now search for main.mix in folders called allied, soviet, counterstrike and aftermath
when trying to detect a CD before performing an actual CD lookup.

If one folder is provided it will behave as if -CD was used and searching for actual physical CDs
will be disabled. To allow disk switching for access to all content all folders must be provided.