Implement SDL2 hardware cursor with scaling
This is a naive implementation for animated cursors where the
cursor is being re-created every time an animation frame is
executed.

Software cursor is also reworked for SDL2 and is now being
rendered directly during flip from the same data.

Should still be plenty performant and responsive in both modes.