SDL2 fullscreen toggle and improved cursor locking
Alt-Enter now toggles between configured windowed size and
fullscreen. The default configuration uses native fullscreen and
original window size.

Cursor locking has been moved to the video driver as it makes
more sense as the video driver also natively knows the scaling
factor which needs to be updated when changing between fullscreen
and windowed.

Games now request locking the cursor only when no in-game dialog
is open and always release it in any other state. No locking is
requested until the main game loop is running.

When toggling between windowed and fullscreen the game requested
state is re-evaluated and applied if necessary. This means that
the cursor will always be locked when going fullscreen and
released in windowed mode if you're in any of the menus.

The SDL2 video driver honors the game locking requests in windowed
mode but ignores in fullscreen and always relocks the cursor
regardless what the request was.

The DirectDraw video driver ignores all locking requests and does
a one time ClipCursor call on video mode set and release it when
the mode is reset. This is likely broken in some way but since we
are phasing out DirectDraw and CnC-DDraw is likely used for testing
I'm not very interested in making this more robust.