Update README and refresh CMakeLists
Set Vanilla as the default and only target by default, reorder
README so that it is preferred over Remaster.

Windows builds now default to SDL2 and OpenAL. Networking is now
disabled on CI until we make it work because the CI artifacts are
used for testing and they wouldn't represent the testable state.

Dropped Wine checking and setting the emulator because the test
was wrong and we don't need to run tests with Wine anymore on
Linux.

Use CMake feature summary for displaying enabled/disabled
features.

These changes make all builds equal feature wise by default.