Change SDL2 default configuration
 - use boxing by default
 - use raw mouse input by default
 - use software cursor by default

Boxing is likely the expected default how the game should function
when going fullscreen aka not stretch the image. It also requires
software cursor to be used as hardware cursor would be off so that
needs to change as well.

While boxing does not strictly require raw input as the mouse
position is automatically scaled to match the visible area it does
create a scaling artifact where your horizontal or vertical cursor
speed is slower than it should be.

Even with raw input enabled it doesn't affect cursor grabbing in
windowed mode where when not in-game it will act like raw input
was not enabled and is free to exit the window.