Remove legacy Resolution option
This was used to create a larger virtual screen with the game
graphics buffers. It seems that this option was not fully honored
anymore and caused artifacts in the lower portion of the screen.

We now support automatic boxing to any size on SDL2 side of things
which will also adjust mouse position correctly which didn't work
with the Resolution option as it expected the cursor to be
confined in the correct area which was not possible with SDL2.

The ScreenHeight global defaults to the expected internal height
in standalone and remaster so all adjustments to it has been
removed until we have proper hires.

Legacy DirectDraw build is the most affected by this and makes it
impossible to run it in 640x480.