Do not crash if critical file is not found
If a critical file (such as temperat.pal) is not found in the MIX
files, the game crashes. This commit fixes this by checking if
opened file is not NULL.

The original crash reason was because an incorrect version of Is_Open
was being called. This was fixed by explicit calling the correct
function.

We also change DBG_FATAL_COMMON to call ShowSimpleMessageBox.

Closes #486