Move SDL messagebox to debugstring.cpp
Previously, we called SDL_SimpleMessageBox on debugstring.h, which
could add this undefined symbol to the compilation unit unless
SDL_messagebox.h was already defined. This commit fixes this by
wrapping this call in a function on debugstring.cpp

We also switched to MessageBoxA if the build targets Windows, once
it is always present on it (at least Win95).

We also try to call the correct version of class functions, but
we are not sure if we covered everything. The crash about incorrect
temperat.pal is being correctly displayed.