Fix unaligned accesses and tweak some stuff for non-standard system implementations
Systems running old ARM CPUs (ARM7, ARM9) has issues when accessing
unaligned addresses, resulting in swapped values. This commit fixes most
(if not all) of the issues regarding Tiberian Dawn.

There is still unaligned access warnings on BigShpBuffer, but that is not
an issue because althrough it reads/writes wrong value in memory, this is
correctly swapped in the CPU.

This commit also fixes issues with obscure filesystem implementations,
like fseek being slow when it should return imediatelly and not finding
upper-case files on FAT32.

This also fixes a leak on Tiberian Dawn scorescreen.

Signed-off-by: Giuliano Belinassi <giuliano.belinassi@gmail.com>