Endian fixes for saving and loading games.
The savegame code stores entire objects to disk almost verbatim, only converting pointers into indices or a coded format before saving, and restoring them back to pointers after loading. It's this decoding that's broken on big endian, as the code takes a pointer to the pointer being decoded, but casts that pointer to a pointer to a type that's smaller than a pointer, and then dereferences the pointer to the smaller type, and thus on big endian only reads the upper byte of the pointer being decoded, where as on little endian it reads the lower byte of the pointer being decoded, which contains the correct value.