Endian/alignment fixes for unions in Red Alert.
Various unions in Red Alert mix variable types of different width, but expects to be able to write a value to a variable of one width, and then read the same value back from a variable of a different width. This works fine on little endian, but not on big endian, so this change adds padding on big endian before the variables with a smaller width.