Replace strlen upper bound with checks for '\0'
Strings in C are always terminated by the '\0' character. We can
explore this to avoid sweeping through the string twice.

Constructs like `for (i = 0; i < strlen(str); i++)` are even worse, as
it will call strlen **for each character of the string**, which is
O(n²) instead of O(n). Replace those cases as well.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>